[Android] Enable auto-generating js-stub based on annotations
This commit contains the feature implementation and a test case.
Currently, only object extension is supported, without function or
constructor extensions. This means, the entry point of the extension
exposed to JavaScript will be an object instance not a function.

Following detail features are supported.
1. Java property binding to JavaScript property, writable/readonly.
Property accessing is sync from JavaScript. Java implementation
of the extension is responsible to notify JavaScript to update
exposed properties.
2. Java Method binding to JavaScript member function of the extension
object. Callback and Promise are supported to the member functions.
3. Simple event model. Object extensions can be EventTargets to response
events triggered by Java side, and of cause the JavaScript side.